Add stylized tooltips to mapbox.js map
This commit does not clearly show what was needed to get tooltips to work.  Specifically, the project.yml file needed to be manually updated to enable UTFGrid interactivity. This then needed to be saved and uploaded to mapbox via the standalone Mapbox Studio Classic app.
References:
**https://www.mapbox.com/help/studio-classic-styles/#add-pop-ups
**https://github.com/mapbox/utfgrid-spec/blob/master/1.3/interaction.md#template
** http://mustache.github.io/
**http://mustache.github.io/mustache.5.html
** https://www.mapbox.com/mapbox.js/api/v3.1.1/l-mapbox-gridcontrol/#section-gridcontrol-hide
**https://stackoverflow.com/questions/23464370/mapbox-js-tooltips-position-over-marker-but-dont-follow-mouse

As a sidenote, the reason why this somewhat odd approach (convoluted if nothing else) was needed, is that the tooltips are being generated from the data loaded in the Mapbox classic style (i.e. alisterfx.a4f587a5). Otherwise, the data would need to be loaded directly via GeoJSON.